change color popup UI

+++ Assets/Scripts/Ternbase/CardTernBase/HealthSystem.cs
@@ -1,33 +1,32 @@
 using System;
-using TMPro;
 using UnityEngine;
 public class HealthSystem 
 {
     public event EventHandler OnHealthChanged;
-    private int health;
-    private int healthMax;
+    private float health;
+    private float healthMax;
 
     public HealthSystem(int healthMax){
         this.healthMax = healthMax;
         this.health = healthMax;
     }
 
-    public int GetHealth(){
+    public float GetHealth(){
         return health;
     }
 
     public float GetHealthPercent(){
-        return (float)health/healthMax;
+        return health/healthMax;
     }
 
-    public void Damage(int damageAmount){
+    public void Damage(float damageAmount){
         health -= damageAmount;
         Debug.Log("Damaged " + damageAmount);
         if(health <= 0) health = 0;
         if(OnHealthChanged != null) OnHealthChanged(this, EventArgs.Empty);
     }
 
-    public void Heal(int healAmount){
+    public void Heal(float healAmount){
         health += healAmount;
         Debug.Log("heal +" + healAmount);
         if(health >= healthMax) health = healthMax;
